AD-3 Refactoring
