+++ app/src/main/java/araikovich/inc/araikovichdemo/data/mappers/GitHubRepoEntityToModelMapper.kt
@@ -1,7 +1,7 @@
 package araikovich.inc.araikovichdemo.data.mappers
 
 import araikovich.inc.araikovichdemo.data.models.entities.GitHubRepoEntity
-import araikovich.inc.araikovichdemo.data.models.ui_models.GitHubRepoModel
+import araikovich.inc.araikovichdemo.ui.main_screen.models.GitHubRepoModel
 
 class GitHubRepoEntityToModelMapper : Mapper<GitHubRepoEntity, GitHubRepoModel> {
