+++ app/src/main/java/araikovich/inc/araikovichdemo/datasource/utils/Result.kt
@@ -1,12 +1,7 @@
 package araikovich.inc.araikovichdemo.datasource.utils
 
-import java.net.HttpURLConnection
-
 sealed class Result<out T> {
-    data class Success<out T>(val code: Int, val data: T?) : Result<T>() {
-        fun isCodeSuccess(): Boolean =
-            code == HttpURLConnection.HTTP_OK || code == HttpURLConnection.HTTP_CREATED
-    }
+    data class Success<out T>(val code: Int, val data: T?) : Result<T>()
 
     data class Error(
         val exception: Exception,
