+++ app/src/main/res/layout/activity_main.xml
@@ -16,7 +16,7 @@
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appbar"
         android:layout_width="match_parent"
-        android:layout_height="269dp"
+        android:layout_height="@dimen/app_bar_height"
         android:background="@color/color_transparent"
         android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
         app:elevation="0dp">
@@ -39,9 +39,9 @@
 
                 <androidx.appcompat.widget.AppCompatImageView
                     android:id="@+id/iv_araikovich"
-                    android:layout_width="150dp"
-                    android:layout_height="150dp"
-                    android:layout_marginTop="14dp"
+                    android:layout_width="@dimen/main_image_width"
+                    android:layout_height="@dimen/main_image_height"
+                    android:layout_marginTop="@dimen/main_image_margin_top"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toTopOf="parent" />
@@ -70,7 +70,7 @@
                         android:fontFamily="@font/roboto_bold"
                         android:text="@string/araikovich_demo"
                         android:textColor="@color/white"
-                        android:textSize="18dp"
+                        android:textSize="@dimen/tv_title_text_size"
                         app:layout_constraintBottom_toBottomOf="parent"
                         app:layout_constraintEnd_toEndOf="parent"
                         app:layout_constraintStart_toStartOf="parent"
@@ -78,10 +78,10 @@
 
                     <androidx.appcompat.widget.AppCompatImageView
                         android:id="@+id/info"
-                        android:layout_width="60dp"
-                        android:layout_height="30dp"
-                        android:paddingStart="15dp"
-                        android:paddingEnd="15dp"
+                        android:layout_width="@dimen/info_width"
+                        android:layout_height="@dimen/info_height"
+                        android:paddingStart="@dimen/info_pagging_start"
+                        android:paddingEnd="@dimen/info_paging_end"
                         app:layout_constraintBottom_toBottomOf="parent"
                         app:layout_constraintEnd_toEndOf="parent"
                         app:layout_constraintTop_toTopOf="parent"
@@ -104,9 +104,9 @@
             android:id="@+id/rv_items"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:layout_marginStart="15dp"
-            android:layout_marginTop="9dp"
-            android:layout_marginEnd="15dp"
+            android:layout_marginStart="@dimen/rv_items_magin_start"
+            android:layout_marginTop="@dimen/rv_items_margin_top"
+            android:layout_marginEnd="@dimen/rv_items_margin_end"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
@@ -114,8 +114,8 @@
 
         <ProgressBar
             android:id="@+id/progressBar"
-            android:layout_width="70dp"
-            android:layout_height="70dp"
+            android:layout_width="@dimen/progress_bar_width"
+            android:layout_height="@dimen/progress_bar_height"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
