+++ app/src/main/res/layout/view_holder_repo_item.xml
@@ -4,24 +4,24 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:background="@drawable/repo_item_background"
-    android:minHeight="70dp">
+    android:minHeight="@dimen/item_container_min_height">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/iv_avatar"
-        android:layout_width="40dp"
-        android:layout_height="40dp"
-        android:layout_marginStart="15dp"
-        android:layout_marginTop="15dp"
-        android:layout_marginBottom="15dp"
+        android:layout_width="@dimen/item_iv_avatar_width"
+        android:layout_height="@dimen/item_iv_avatar_height"
+        android:layout_marginStart="@dimen/item_iv_avatar_margin_start"
+        android:layout_marginTop="@dimen/item_iv_avatar_margin_top"
+        android:layout_marginBottom="@dimen/item_iv_avatar_margin_bottom"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/iv_favourite"
-        android:layout_width="40dp"
-        android:layout_height="40dp"
-        android:layout_marginTop="15dp"
-        android:layout_marginEnd="30dp"
+        android:layout_width="@dimen/item_iv_favourite_width"
+        android:layout_height="@dimen/item_iv_favourite_height"
+        android:layout_marginTop="@dimen/item_iv_favourite_margin_top"
+        android:layout_marginEnd="@dimen/item_iv_favourite_margin_end"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -29,12 +29,12 @@
         android:id="@+id/tv_name"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="15dp"
-        android:layout_marginEnd="15dp"
+        android:layout_marginStart="@dimen/item_tv_name_margin_start"
+        android:layout_marginEnd="@dimen/item_tv_name_margin_end"
         android:fontFamily="@font/roboto_bold"
         android:gravity="start"
         android:textColor="@color/white"
-        android:textSize="16dp"
+        android:textSize="@dimen/item_tv_name_text_size"
         app:layout_constraintBottom_toBottomOf="@+id/iv_avatar"
         app:layout_constraintEnd_toStartOf="@+id/iv_favourite"
         app:layout_constraintStart_toEndOf="@+id/iv_avatar"
@@ -44,14 +44,14 @@
         android:id="@+id/tv_description"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="15dp"
-        android:layout_marginTop="15dp"
-        android:layout_marginEnd="15dp"
-        android:layout_marginBottom="15dp"
+        android:layout_marginStart="@dimen/item_tv_description_margin_start"
+        android:layout_marginTop="@dimen/item_tv_description_margin_top"
+        android:layout_marginEnd="@dimen/item_tv_description_margin_end"
+        android:layout_marginBottom="@dimen/item_tv_description_margin_bottom"
         android:fontFamily="@font/roboto_regular"
         android:gravity="start"
         android:textColor="@color/white"
-        android:textSize="13dp"
+        android:textSize="@dimen/item_tv_description_text_size"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toStartOf="@+id/iv_favourite"
         app:layout_constraintStart_toEndOf="@+id/iv_avatar"
