+++ app/src/main/res/values/dimens.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="app_bar_height">269dp</dimen>
+    <dimen name="main_image_width">150dp</dimen>
+    <dimen name="main_image_height">150dp</dimen>
+    <dimen name="main_image_margin_top">14dp</dimen>
+    <dimen name="tv_title_text_size">18dp</dimen>
+    <dimen name="info_width">60dp</dimen>
+    <dimen name="info_height">30dp</dimen>
+    <dimen name="info_pagging_start">15dp</dimen>
+    <dimen name="info_paging_end">15dp</dimen>
+    <dimen name="rv_items_magin_start">15dp</dimen>
+    <dimen name="rv_items_margin_top">9dp</dimen>
+    <dimen name="rv_items_margin_end">15dp</dimen>
+    <dimen name="progress_bar_width">70dp</dimen>
+    <dimen name="progress_bar_height">70dp</dimen>
+    <dimen name="item_iv_avatar_width">40dp</dimen>
+    <dimen name="item_iv_avatar_height">40dp</dimen>
+    <dimen name="item_iv_avatar_margin_start">15dp</dimen>
+    <dimen name="item_iv_avatar_margin_top">15dp</dimen>
+    <dimen name="item_iv_avatar_margin_bottom">15dp</dimen>
+    <dimen name="item_iv_favourite_width">40dp</dimen>
+    <dimen name="item_iv_favourite_height">40dp</dimen>
+    <dimen name="item_iv_favourite_margin_top">15dp</dimen>
+    <dimen name="item_iv_favourite_margin_end">30dp</dimen>
+    <dimen name="item_tv_name_margin_start">15dp</dimen>
+    <dimen name="item_tv_name_margin_end">15dp</dimen>
+    <dimen name="item_tv_name_text_size">16dp</dimen>
+    <dimen name="item_tv_description_margin_start">15dp</dimen>
+    <dimen name="item_tv_description_margin_top">15dp</dimen>
+    <dimen name="item_tv_description_margin_end">15dp</dimen>
+    <dimen name="item_tv_description_margin_bottom">15dp</dimen>
+    <dimen name="item_tv_description_text_size">13dp</dimen>
+    <dimen name="item_container_min_height">70dp</dimen>
+</resources>
